# frontend/api-swagger.json
{"openapi":"3.0.1","info":{"title":"HospitalApi","version":"v1"},"paths":{"/api/Apointment":{"post":{"tags":["Apointment"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddAppointmentToDoctorRequest"}},"text/json":{"schema":{"$ref":"#/components/schemas/AddAppointmentToDoctorRequest"}},"application/*+json":{"schema":{"$ref":"#/components/schemas/AddAppointmentToDoctorRequest"}}}},"responses":{"200":{"description":"OK"}}}},"/api/Apointment/me/{count}":{"get":{"tags":["Apointment"],"parameters":[{"name":"count","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/Apointment/{doctorLogin}/{count}":{"get":{"tags":["Apointment"],"parameters":[{"name":"doctorLogin","in":"path","required":true,"schema":{"type":"string"}},{"name":"count","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/Apointment/me":{"post":{"tags":["Apointment"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppointmentAddRequest"}},"text/json":{"schema":{"$ref":"#/components/schemas/AppointmentAddRequest"}},"application/*+json":{"schema":{"$ref":"#/components/schemas/AppointmentAddRequest"}}}},"responses":{"200":{"description":"OK"}}}},"/api/Apointment/{appointmentId}":{"delete":{"tags":["Apointment"],"parameters":[{"name":"appointmentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/Auth/login":{"post":{"tags":["Auth"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialsReq"}},"text/json":{"schema":{"$ref":"#/components/schemas/CredentialsReq"}},"application/*+json":{"schema":{"$ref":"#/components/schemas/CredentialsReq"}}}},"responses":{"200":{"description":"OK"}}}},"/api/Auth/register":{"post":{"tags":["Auth"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialsReq"}},"text/json":{"schema":{"$ref":"#/components/schemas/CredentialsReq"}},"application/*+json":{"schema":{"$ref":"#/components/schemas/CredentialsReq"}}}},"responses":{"200":{"description":"OK"}}}},"/api/Person/register":{"post":{"tags":["Person"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonRequest"}},"text/json":{"schema":{"$ref":"#/components/schemas/PersonRequest"}},"application/*+json":{"schema":{"$ref":"#/components/schemas/PersonRequest"}}}},"responses":{"200":{"description":"OK"}}}},"/api/Person/me":{"get":{"tags":["Person"],"responses":{"200":{"description":"OK"}}}},"/api/Person/role":{"post":{"tags":["Person"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetRoleRequest"}},"text/json":{"schema":{"$ref":"#/components/schemas/SetRoleRequest"}},"application/*+json":{"schema":{"$ref":"#/components/schemas/SetRoleRequest"}}}},"responses":{"200":{"description":"OK"}}}},"/api/Temp/get/{id}":{"get":{"tags":["Temp"],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/Temp/create":{"post":{"tags":["Temp"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonRequest"}},"text/json":{"schema":{"$ref":"#/components/schemas/PersonRequest"}},"application/*+json":{"schema":{"$ref":"#/components/schemas/PersonRequest"}}}},"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"AddAppointmentToDoctorRequest":{"type":"object","properties":{"doctorLogin":{"type":"string","nullable":true},"patientLogin":{"type":"string","nullable":true},"slot":{"$ref":"#/components/schemas/Slot"}},"additionalProperties":false},"AppointmentAddRequest":{"type":"object","properties":{"patientLogin":{"type":"string","nullable":true},"slot":{"$ref":"#/components/schemas/Slot"}},"additionalProperties":false},"CredentialsReq":{"type":"object","properties":{"login":{"type":"string","nullable":true},"password":{"type":"string","nullable":true}},"additionalProperties":false},"PersonRequest":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string","nullable":true},"surname":{"type":"string","nullable":true},"pesel":{"type":"string","nullable":true}},"additionalProperties":false},"PersonRole":{"enum":[0,1,2],"type":"integer","format":"int32"},"SetRoleRequest":{"type":"object","properties":{"personId":{"type":"string","format":"uuid"},"role":{"$ref":"#/components/schemas/PersonRole"}},"additionalProperties":false},"Slot":{"type":"object","properties":{"start":{"type":"string","format":"date-time"},"end":{"type":"string","format":"date-time"}},"additionalProperties":false}},"securitySchemes":{"Bearer":{"type":"http","description":"Put **_ONLY_** your JWT Bearer token below.","scheme":"bearer","bearerFormat":"JWT"}}},"security":[{"Bearer":[]}]}